Repackage objects in the Units directory
