--- tv/src/main/java/com/californiadreamshostel/officetv/UNIT/CONVERTER/Convertor.java
@@ -1,7 +1,0 @@
-package com.californiadreamshostel.officetv.UNIT.CONVERTER;
-
-public interface Convertor {
-
-    double getResult();
-    Convertor bind(double from);
-}
