--- tv/src/main/java/com/californiadreamshostel/officetv/UNIT/CONVERTER/TEMPERATURE/CelsuisToFarenheitConverter.java
@@ -1,13 +1,0 @@
-package com.californiadreamshostel.officetv.UNIT.CONVERTER.TEMPERATURE;
-
-import com.californiadreamshostel.officetv.UNIT.CONVERTER.BaseConverter;
-import com.californiadreamshostel.officetv.UNIT.CONVERTER.CONVERSIONALGORITHMS.CelsuisToFarenheitAlgorithm;
-import com.californiadreamshostel.officetv.UNIT.CONVERTER.CONVERSIONALGORITHMS.ConversionAlgorithm;
-
-public class CelsuisToFarenheitConverter extends BaseConverter{
-
-    @Override
-    protected ConversionAlgorithm getConversionAlgorithm() {
-        return new CelsuisToFarenheitAlgorithm();
-    }
-}
