--- tv/src/main/java/com/californiadreamshostel/officetv/UNIT/CONVERTER/TEMPERATURE/FarenheitToCelsuisConverter.java
@@ -1,14 +1,0 @@
-package com.californiadreamshostel.officetv.UNIT.CONVERTER.TEMPERATURE;
-
-import com.californiadreamshostel.officetv.UNIT.CONVERTER.BaseConverter;
-import com.californiadreamshostel.officetv.UNIT.CONVERTER.CONVERSIONALGORITHMS.FarenheitToCelsuisAlgorithm;
-import com.californiadreamshostel.officetv.UNIT.CONVERTER.CONVERSIONALGORITHMS.ConversionAlgorithm;
-
-public class FarenheitToCelsuisConverter extends BaseConverter {
-
-    @Override
-    protected ConversionAlgorithm getConversionAlgorithm() {
-        return new FarenheitToCelsuisAlgorithm();
-    }
-
-}
