--- tv/src/main/java/com/californiadreamshostel/officetv/UNIT/UNITS/Ft.java
@@ -1,9 +1,0 @@
-package com.californiadreamshostel.officetv.UNIT.UNITS;
-
-public class Ft extends Unit {
-    public static final String TYPE = "FT";
-
-    {
-        UNIT_TYPE = TYPE;
-    }
-}
