--- tv/src/main/java/com/californiadreamshostel/officetv/UNIT/UNITS/Km.java
@@ -1,9 +1,0 @@
-package com.californiadreamshostel.officetv.UNIT.UNITS;
-
-public class Km extends Unit {
-    public static final String TYPE = "KM";
-
-    {
-        UNIT_TYPE = TYPE;
-    }
-}
