--- tv/src/main/java/com/californiadreamshostel/officetv/UNIT/UNITS/M.java
@@ -1,9 +1,0 @@
-package com.californiadreamshostel.officetv.UNIT.UNITS;
-
-public class M extends Unit {
-    public static final String TYPE = "M";
-
-    {
-        UNIT_TYPE = TYPE;
-    }
-}
