+++ tv/src/main/java/com/californiadreamshostel/officetv/UNIT/UNITS/SIZE/Ft.java
@@ -1,0 +1,11 @@
+package com.californiadreamshostel.officetv.UNIT.UNITS.DISTANCE;
+
+import com.californiadreamshostel.officetv.UNIT.UNITS.Unit;
+
+public class Ft extends Unit {
+    public static final String TYPE = "FT";
+
+    {
+        UNIT_TYPE = TYPE;
+    }
+}
