+++ tv/src/main/java/com/californiadreamshostel/officetv/UNIT/UNITS/SIZE/Km.java
@@ -1,0 +1,11 @@
+package com.californiadreamshostel.officetv.UNIT.UNITS.DISTANCE;
+
+import com.californiadreamshostel.officetv.UNIT.UNITS.Unit;
+
+public class Km extends Unit {
+    public static final String TYPE = "KM";
+
+    {
+        UNIT_TYPE = TYPE;
+    }
+}
