+++ tv/src/main/java/com/californiadreamshostel/officetv/UNIT/UNITS/SIZE/M.java
@@ -1,0 +1,11 @@
+package com.californiadreamshostel.officetv.UNIT.UNITS.DISTANCE;
+
+import com.californiadreamshostel.officetv.UNIT.UNITS.Unit;
+
+public class M extends Unit {
+    public static final String TYPE = "M";
+
+    {
+        UNIT_TYPE = TYPE;
+    }
+}
