--- tv/src/main/java/com/californiadreamshostel/officetv/UNIT/UNITS/UnitType.java
@@ -1,5 +1,0 @@
-package com.californiadreamshostel.officetv.UNIT.UNITS;
-
-public interface UnitType {
-    String getType();
-}
